Move GCAPI example stuff to testmod
