+++ src/test/java/net/modificationstation/sltest/gcapi/ExampleConfig.java
@@ -1,4 +1,4 @@
-package net.modificationstation.stationapi.impl.config.example;
+package net.modificationstation.sltest.gcapi;
 
 import net.modificationstation.stationapi.api.config.ConfigRoot;
