+++ src/test/resources/fabric.mod.json
@@ -46,6 +46,19 @@
     ],
     "preLaunch": [
       "net.modificationstation.sltest.PreLaunchTest"
+    ],
+    "gcapi": [
+      "net.modificationstation.sltest.gcapi.ExampleConfig"
+    ],
+    "gcapi:factory_provider": [
+      "net.modificationstation.stationapi.impl.config.factory.DefaultFactoryProvider",
+      "net.modificationstation.sltest.gcapi.ExampleConfigEnumFactories"
+    ],
+    "gcapi:postload": [
+      "net.modificationstation.sltest.gcapi.ExampleEntryPointListeners"
+    ],
+    "gcapi:presave": [
+      "net.modificationstation.sltest.gcapi.ExampleEntryPointListeners"
     ]
   },
